Move HStar2 impl out to its own file
